requirements.txt
aiogram>=3.4,<4.0
aiofiles==23.2.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
paramiko==3.4.0
cryptography==42.0.5
aiosqlite==0.20.0
cachetools==5.3.2
qrcode[pil]==7.4.2
pillow==10.2.0